requirements.txt
appdirs==1.4.4
asgiref==3.4.1
aspy.refactor-imports==2.2.0
attrs==21.2.0
autoflake==1.4
bandit==1.7.0
black==21.6b0
cached-property==1.5.2
certifi==2021.5.30
cfgv==3.3.0
chardet==4.0.0
click==8.0.1
crypto==1.4.1
distlib==0.3.2
Django==3.2.5
eradicate==2.0.0
filelock==3.0.12
flake8==3.9.2
flake8-annotations==2.6.2
flake8-annotations-coverage==0.0.5
flake8-bandit==2.1.2
flake8-broken-line==0.3.0
flake8-bugbear==21.4.3
flake8-builtins==1.5.3
flake8-commas==2.0.0
flake8-comprehensions==3.5.0
flake8-debugger==4.0.0
flake8-eradicate==1.1.0
flake8-executable==2.1.1
flake8-fixme==1.1.1
flake8-future-import==0.4.6
flake8-mutable==1.2.0
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-pyi==20.10.0
flake8-pytest==1.3
flake8-pytest-style==1.5.0
flake8-string-format==0.3.0
flake8-todo==0.7
flake8-unused-arguments==0.0.6
gitdb==4.0.7
GitPython==3.1.18
identify==2.2.10
idna==2.10
iniconfig==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
Naked==0.1.31
nodeenv==1.6.0
packaging==20.9
pathspec==0.8.1
pbr==5.6.0
pluggy==0.13.1
pre-commit==2.13.0
psycopg2-binary==2.9.1
py==1.10.0
pycodestyle==2.7.0
pycryptodome==3.10.1
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
python-dotenv==0.18.0
pytz==2021.1
pyupgrade==2.19.4
PyYAML==5.4.1
regex==2021.7.1
reorder-python-imports==2.5.0
requests==2.25.1
shellescape==3.8.1
shortuuid==1.0.1
six==1.16.0
smmap==4.0.0
sqlparse==0.4.1
stevedore==3.3.0
tokenize-rt==4.1.0
toml==0.10.2
urllib3==1.26.6
virtualenv==20.4.7
yesqa==1.2.3
